[#55] made it impossible to delete admin in backend

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/service/impl/OperatorServiceImpl.java
@@ -5,6 +5,7 @@ import at.ac.tuwien.sepm.groupphase.backend.config.EncoderConfig;
 import at.ac.tuwien.sepm.groupphase.backend.entity.Operator;
 import at.ac.tuwien.sepm.groupphase.backend.entity.Permissions;
 import at.ac.tuwien.sepm.groupphase.backend.exception.NotFoundException;
+import at.ac.tuwien.sepm.groupphase.backend.exception.ValidationException;
 import at.ac.tuwien.sepm.groupphase.backend.repository.OperatorRepository;
 import at.ac.tuwien.sepm.groupphase.backend.service.OperatorService;
 import at.ac.tuwien.sepm.groupphase.backend.util.OperatorSpecifications;
@@ -138,8 +139,11 @@ public class OperatorServiceImpl implements OperatorService {
     @Override
     public void delete(Long id) {
         LOGGER.trace("delete({})", id);
-        operatorRepository.findById(id)
+        Operator operator = operatorRepository.findById(id)
             .orElseThrow(() -> new NotFoundException("Could not find operator that should be deleted!"));
+        if (operator.getPermissions().equals(Permissions.admin)) {
+            throw new ValidationException("Cannot delete an Admin");
+        }
         operatorRepository.deleteById(id);
     }
